Cards de todos os produtos

+++ inrise/src/components/productCards/RamCard.vue
@@ -1,10 +1,12 @@
 <template>
-  <ProductCard :product="ramProduct" :formatCurrency="formatCurrency"
-  @delete-product="handleDeleteRam">
+  <ProductCard 
+    :product="product" 
+    :formatCurrency="formatCurrency" 
+    @delete-product="handleDeleteRam">
     <template #default="{ product }">
-      <p>Socket: {{ product.socket }}</p>
-      <p>Frequência: {{ product.frequency }} MHz</p>
-      <p>Capacidade: {{ product.capacity }} GB</p>
+      <p>Socket: {{ product.socket || 'Socket não disponível' }}</p>
+      <p>Frequência: {{ product.frequency || 0 }} MHz</p>
+      <p>Capacidade: {{ product.capacity || 0 }} GB</p>
     </template>
   </ProductCard>
 </template>
@@ -18,28 +20,28 @@ export default {
   components: {
     ProductCard,
   },
-  data() {
-    return {
-      ramProduct: {
-        id: 1,
-        name: '',
-        socket: '',
-        frequency: 0,
-        capacity: 0,
-        price: { finalPrice: 0 },
-      },
-    };
+  props: {
+    product: {
+      type: Object,
+      required: true,
+    },
   },
   methods: {
     formatCurrency(value) {
       return `R$ ${value.toFixed(2)}`;
     },
     async handleDeleteRam(product) {
+      if (!product.id) {
+        alert('Produto sem ID para exclusão');
+        return;
+      }
+
       try {
         await deleteRam(product.id);
         alert('Memória RAM excluída com sucesso!');
       } catch (error) {
         alert('Erro ao excluir a memória RAM');
+        console.error(error);
       }
     },
   },
